Extract throwIfDocumentInvalid
