+++ lib/index.js
@@ -2,6 +2,7 @@ import getPosts from './getPosts.js';
 import groupBy from './groupBy.js';
 import { ElementInsertionPosition } from './insertNodeRelativeToElement.js';
 import insertSeriesInfo from './insertSeriesInfo.js';
+import throwIfDocumentInvalid from './throwIfDocumentInvalid.js';
 
 export { ElementInsertionPosition } from './insertNodeRelativeToElement.js';
 
@@ -99,19 +100,6 @@ export function SeriesDisplay(ghostContentApi) {
     function throwIfDisplayOptionsInvalid(/** @type DisplaySeriesInfoOptions */ options) {
         if (options.insertions == undefined || !Array.isArray(options.insertions)) {
             throw new TypeError('option.insertions is required and must be an array.');
-        }
-    }
-
-    function throwIfDocumentInvalid(document) {
-        if (document == undefined) {
-            throw new TypeError('document must be provided.');
-        }
-
-        if (typeof Document === 'undefined') return;
-
-        const isDocument = (value) => (value instanceof Document || value instanceof HTMLDocument);
-        if (Document !== undefined && !isDocument(document)) {
-            throw new TypeError('document must be an instance of Document (or HTMLDocument).');
         }
     }
